move AsyncToSerialHelper class to functions
This avoids the possibility of creating a circular reference that the gc
cannot collect.
